new/requirements.txt
opencv_python==4.5.2.52
numpy==1.20.1
Pillow==8.3.1
PyQt5==5.15.4
